`SymbolDataCandle`'s `Volume` should be `float64` in Go language files

+++ defaults/languages/go-language/trading/data.go
@@ -23,7 +23,7 @@ type SymbolDataCandle struct {
 	High      float64
 	Low       float64
 	Close     float64
-	Volume    uint64
+	Volume    float64
 	Timestamp uint64
 	Status    string
 }
